restructure project files
